+++ NEWS.md
@@ -1,3 +1,5 @@
+# utsf (development version)
+
 # utsf 1.1.0
 
 * Improvements in estimation of forecast accuracy with rolling origin evaluation.